% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esti_polyserial.R
\name{esti_polyserial}
\alias{esti_polyserial}
\title{Polyserial Correlation}
\usage{
esti_polyserial(X, maxn = 100, e = 1e-08)
}
\arguments{
\item{X}{a matrix or dataframe contains two polyserial variable.}

\item{maxn}{the maximum iterations times.}

\item{e}{the maximum tolerance of convergence.}
}
\value{
\item{rho}{estimated value of polyserial correlation coefficient.}

\item{std}{standard deviation of rho.}

\item{iter}{times of iteration convergence.}

\item{Ex,Ey}{the support point of regression model.}
}
\description{
Estimate the polyserial correlation coefficient.
}
\examples{
X = gen_polyseries(1000,0.5,-1:1)
result = esti_polyserial(X)
result
}
\seealso{
\link{esti_polychoric}
}
