% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_polychoric.R
\name{gen_polychoric}
\alias{gen_polychoric}
\title{Generate Polychoric Sample}
\usage{
gen_polychoric(n, rho, a, b)
}
\arguments{
\item{n}{sample size.}

\item{rho}{correlation coefficient.}

\item{a}{the cutoff points array.}

\item{b}{the cutoff points array.}
}
\value{
Polychoric sample with size n(in a 2*n matrix).
}
\description{
Generate polychoric sample with hidden distribution: binormal with correlation coefficient rho.
}
\examples{
gen_polychoric(100,0.5,-1:1,1:2)
}
\seealso{
\link{gen_polyseries}
\link{gen_rho}
}
