% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_polyseries.R
\name{gen_polyseries}
\alias{gen_polyseries}
\title{Generate Polyseries Sample}
\usage{
gen_polyseries(n, rho, a)
}
\arguments{
\item{n}{sample size.}

\item{rho}{correlation coefficient.}

\item{a}{the cutoff points array.}
}
\value{
Polyseries sample with size n(in a 2*n matrix).
}
\description{
Generate polyseries sample with hidden distribution: binormal with correlation coefficient rho.
}
\examples{
gen_polyseries(100,0.5,-1:1)
}
\seealso{
\link{gen_rho}
\link{gen_polychoric}
}
