% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rho.R
\name{gen_rho}
\alias{gen_rho}
\title{Generate Specific Binormal Distribution}
\usage{
gen_rho(n, rho)
}
\arguments{
\item{n}{sample size.}

\item{rho}{correlation coefficient.}
}
\value{
Binormal random number with length n(in a 2*n matrix).
}
\description{
Generate random number of binormal distribution with 0 mean unit variance and correlation coefficient rho.
}
\examples{
gen_rho(100,0.5)
}
\seealso{
\link{gen_polyseries}
\link{gen_polychoric}
}
