% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb.R
\name{mb}
\alias{mb}
\title{Mean Bias}
\usage{
mb(rhohat, rho)
}
\arguments{
\item{rhohat}{an array of estimators of rho.}

\item{rho}{the true value of rho.}
}
\value{
the mean bias of rhohat array.
}
\description{
Calculate the MB of an array of estimates relative to the true value.
}
\examples{
rho = 0.5
rhohat = 0.5 + rnorm(10)
mb(rohat,rho)
}
\seealso{
\link{mrb}
\link{rmse}
}
