% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrb.R
\name{mrb}
\alias{mrb}
\title{Mean Relative Bias}
\usage{
mrb(rhohat, rho)
}
\arguments{
\item{rhohat}{an array of estimators of rho.}

\item{rho}{the true value of rho.}
}
\value{
the mean relative bias of rhohat array.
}
\description{
Calculate the MRB of an array of estimates relative to the true value.
}
\examples{
rho = 0.5
rhohat = 0.5 + rnorm(10)
mrb(rohat,rho)
}
\seealso{
\link{mb}
\link{rmse}
}
