% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_panels1.R
\name{pairs_panels1}
\alias{pairs_panels1}
\title{Draw the SPLOM of several polychoric variables.}
\usage{
pairs_panels1(
  x,
  MLE = FALSE,
  smooth = TRUE,
  scale = FALSE,
  density = TRUE,
  ellipses = TRUE,
  pch = 20,
  jiggle = FALSE,
  factor = 2,
  hist.col = "cyan",
  show.points = TRUE,
  rug = TRUE,
  breaks = "Sturges",
  cex.cor = 1,
  wt = NULL,
  smoother = FALSE,
  stars = FALSE,
  ci = FALSE,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{x}{A dataframe of matrix contains several polychoric variables.}

\item{MLE}{\code{TRUE} for MLE method, \code{FALSE} for IRLS method.}

\item{smooth}{\code{TRUE} for fitting a smooth curve between two variables.}

\item{density}{\code{TRUE} for adding density curves on histograms.}

\item{ellipses}{\code{TRUE} for adding correlation ellipses.}

\item{pch}{The points type.}

\item{jiggle}{\code{TRUE} for adding noise to x.}

\item{factor}{The noise intensity in jiggle.}

\item{hist.col}{The color of histograms.}

\item{show.points}{\code{TRUE} for showing the points.}

\item{rug}{\code{TRUE} for adding rugs to histograms.}

\item{breaks}{Similar to the \item{breaks} parameter in function \item{hist}.}

\item{cex.cor}{The size of coefficients.}

\item{smoother}{\code{TRUE} for adding gradient effect to points.}
}
\value{
SPLOM as described in the description.
}
\description{
Draw a scatter plot of matrices (SPLOM), with bivariate scatter plots below
the diagonal, histograms on the diagonal, and the polychoric correlation
coefficients with standard errors above the diagonal.
Correlation ellipses are drawn in the same graph. The red lines below the
diagonal are the LOESS smoothed lines, fitting a smooth curve between two variables.
}
\examples{
pairs_panels1(CECERS[,1:5])
pairs_panels1(CECERS[,1:5],MLE=T)
}
