% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{PN}
\alias{PN}
\title{Parenteral_nutrition}
\format{
A data frame with 1086 rows and 29 variables:
}
\description{
The Parenteral Nutrition data were collected from 543 patients
of whom 386 were given parenteral nutrition alone, 145 were given enteral and
parenteral nutrition, and 3 were given enteral nutrition only. There are 23
main discrete variables, such as: clinical stages(1-4), dietary status(1-3),
NRS(0-6), PG-SGA-qualitative(1-3), etc.
}
