% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(rhohat, rho)
}
\arguments{
\item{rhohat}{an array of estimators of rho.}

\item{rho}{the true value of rho.}
}
\value{
the root mean squared error of rhohat array.
}
\description{
Calculate the RMSE of an array of estimates relative to the true value.
}
\examples{
rho = 0.5
rhohat = 0.5 + rnorm(10)
rmse(rhohat,rho)
}
\seealso{
\link{mb}
\link{mrb}
}
